@@ -1,23 +1,29 @@
-name: signalr_client
-description: A flutter client for ASP.NET Core SignalR. The client is able to invoke server side hub functions and to receive method invocations issued by the server. 
-version: 0.1.5
-author: Sörnt Poppe <soernt.poppe@nordlichtsoftware.de>
-homepage: https://github.com/soernt/signalr_client
+name: signalr_netcore
+description: A flutter client for ASP.NET Core SignalR. The client is able to invoke server side hub functions and to receive method invocations issued by the server.
+version: 1.3.7
+homepage: https://github.com/sefidgaran/signalr_client
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.12.0 <4.0.0"
 
 dependencies:
-
   flutter:
     sdk: flutter
 
-  logging: ^0.11.3+2
-  w3c_event_source: ^1.3.0
+  logging: ^1.2.0
+  tuple: ^2.0.2
+  web_socket_channel: ^2.4.0
+  sse_channel2: ^0.0.4
+  message_pack_dart: ^2.0.1
+  http: ^1.1.0
 
 dev_dependencies:
+  collection: ^1.15.0
   flutter_test:
     sdk: flutter
+  quiver: ^3.2.1
+  rxdart: ^0.27.7
+  test: ^1.16.2
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
