@@ -1,7 +1,6 @@
 import 'dart:collection';
 
 import 'package:logging/logging.dart';
-import 'package:signalr_client/utils.dart';
 
 import 'errors.dart';
 import 'itransport.dart';
@@ -10,23 +9,23 @@ import 'itransport.dart';
 enum MessageType {
   /// MessageType is not defined.
   Undefined, // = 0,
-  /// Indicates the message is an Invocation message and implements the {@link @aspnet/signalr.InvocationMessage} interface.
+  /// Indicates the message is an Invocation message and implements the {@link @microsoft/signalr.InvocationMessage} interface.
   Invocation, // = 1,
-  /// Indicates the message is a StreamItem message and implements the {@link @aspnet/signalr.StreamItemMessage} interface.
+  /// Indicates the message is a StreamItem message and implements the {@link @microsoft/signalr.StreamItemMessage} interface.
   StreamItem, // = 2,
-  /// Indicates the message is a Completion message and implements the {@link @aspnet/signalr.CompletionMessage} interface.
+  /// Indicates the message is a Completion message and implements the {@link @microsoft/signalr.CompletionMessage} interface.
   Completion, // = 3,
-  /// Indicates the message is a Stream Invocation message and implements the {@link @aspnet/signalr.StreamInvocationMessage} interface.
+  /// Indicates the message is a Stream Invocation message and implements the {@link @microsoft/signalr.StreamInvocationMessage} interface.
   StreamInvocation, // = 4,
-  /// Indicates the message is a Cancel Invocation message and implements the {@link @aspnet/signalr.CancelInvocationMessage} interface.
+  /// Indicates the message is a Cancel Invocation message and implements the {@link @microsoft/signalr.CancelInvocationMessage} interface.
   CancelInvocation, // = 5,
-  /// Indicates the message is a Ping message and implements the {@link @aspnet/signalr.PingMessage} interface.
+  /// Indicates the message is a Ping message and implements the {@link @microsoft/signalr.PingMessage} interface.
   Ping, // = 6,
-  /// Indicates the message is a Close message and implements the {@link @aspnet/signalr.CloseMessage} interface.
+  /// Indicates the message is a Close message and implements the {@link @microsoft/signalr.CloseMessage} interface.
   Close, // = 7,
 }
 
-MessageType parseMessageTypeFromString(int value) {
+MessageType? parseMessageTypeFromString(int? value) {
   if (value == null) {
     return null;
   }
@@ -34,25 +33,18 @@ MessageType parseMessageTypeFromString(int value) {
   switch (value) {
     case 1:
       return MessageType.Invocation;
-      break;
     case 2:
       return MessageType.StreamItem;
-      break;
     case 3:
       return MessageType.Completion;
-      break;
     case 4:
       return MessageType.StreamInvocation;
-      break;
     case 5:
       return MessageType.CancelInvocation;
-      break;
     case 6:
       return MessageType.Ping;
-      break;
     case 7:
       return MessageType.Close;
-      break;
     default:
       throw GeneralError("A MessageType of {value} is not supported.");
   }
@@ -63,11 +55,12 @@ class MessageHeaders {
   static const String AuthorizationHeaderName = "Authorization";
 
   // Properties
-  HashMap<String, String> _headers;
+  HashMap<String, String>? _headers;
 
-  Iterable<String> get names => _headers.keys;
+  Iterable<String> get names => _headers!.keys;
+  HashMap<String, String>? get asMap => _headers;
 
-  bool get isEmtpy => _headers.length == 0;
+  bool get isEmpty => _headers!.length == 0;
 
   // Methods
   MessageHeaders() {
@@ -75,34 +68,42 @@ class MessageHeaders {
   }
 
   /// Gets the header with the specified key.
-  String getHeaderValue(String name) {
-    return _headers[name];
+  String? getHeaderValue(String name) {
+    return _headers![name];
   }
 
   /// Sets the header with the specified key.
   void setHeaderValue(String name, String value) {
-    _headers[name] = value;
+    _headers![name] = value;
   }
 
-  void setAuthorizationHeaderValue(String token) {
-    if (!isStringEmpty(token)) {
-      setHeaderValue(AuthorizationHeaderName, "Bearer $token");
-    } else {
-      removeAuthorizationHeaderValue();
-    }
-  }
-
-  void removeAuthorizationHeaderValue() {
-    if (_headers.containsKey(AuthorizationHeaderName)) {
-      _headers.remove(AuthorizationHeaderName);
-    }
-  }
-
-  /// removes the given header
+  /// Removes the given header
   void removeHeader(String name) {
-    if (_headers.containsKey(name)) {
-      _headers.remove(name);
+    if (_headers!.containsKey(name)) {
+      _headers!.remove(name);
     }
+  }
+
+  /// Concatenate message headers
+  void addMessageHeaders(MessageHeaders? newHeaders) {
+    if (newHeaders != null && !newHeaders.isEmpty) {
+      for (var name in newHeaders.names) {
+        setHeaderValue(name, newHeaders.getHeaderValue(name)!);
+      }
+    }
+  }
+
+  @override
+  String toString() {
+    if (isEmpty) return '{}';
+
+    String str = '';
+    for (var name in names) {
+      if (str.isNotEmpty) str += ', ';
+      str += '{ $name: ${_headers![name]} }';
+    }
+
+    return str;
   }
 }
 
@@ -120,18 +121,18 @@ abstract class HubMessageBase {
 /// Defines properties common to all Hub messages relating to a specific invocation.
 abstract class HubInvocationMessage extends HubMessageBase {
   // Properties
-  /// A {@link @aspnet/signalr.MessageHeaders} dictionary containing headers attached to the message.
+  /// A {@link @microsoft/signalr.MessageHeaders} dictionary containing headers attached to the message.
   final MessageHeaders headers;
 
   ///The ID of the invocation relating to this message.
   ///
   ///This is expected to be present for StreamInvocationMessage and CompletionMessage. It may
   ///be 'undefined' for an InvocationMessage if the sender does not expect a response.
-  final String invocationId;
+  final String? invocationId;
 
   // Methods
   HubInvocationMessage(
-      MessageType messageType, MessageHeaders headers, String invocationId)
+      MessageType messageType, MessageHeaders? headers, String? invocationId)
       : this.headers = headers ?? MessageHeaders(),
         this.invocationId = invocationId,
         super(messageType);
@@ -142,17 +143,29 @@ class InvocationMessage extends HubInvocationMessage {
   // Properites
 
   /// The target method name.
-  final String target;
+  final String? target;
 
   /// The target method arguments.
-  final List<Object> arguments;
+  final List<Object?>? arguments;
+
+  /// The target method's stream IDs.
+  final List<String>? streamIds;
 
   // Methods
-  InvocationMessage(String target, List<Object> arguments,
-      MessageHeaders headers, String invocationId)
+  InvocationMessage(
+      {String? target,
+      List<Object?>? arguments,
+      List<String>? streamIds,
+      MessageHeaders? headers,
+      String? invocationId})
       : this.target = target,
         this.arguments = arguments,
+        this.streamIds = streamIds,
         super(MessageType.Invocation, headers, invocationId);
+  @override
+  String toString() {
+    return 'InvocationMessage - type: ${type.index}, headers: ${headers.toString()}, invocationId: $invocationId, target: $target, arguments: $arguments, streamIds: $streamIds';
+  }
 }
 
 /// A hub message representing a streaming invocation.
@@ -160,17 +173,30 @@ class StreamInvocationMessage extends HubInvocationMessage {
   // Properites
 
   /// The target method name.
-  final String target;
+  final String? target;
 
   /// The target method arguments.
-  final List<Object> arguments;
+  final List<Object>? arguments;
+
+  /// The target method's stream IDs.
+  final List<String>? streamIds;
 
   // Methods
-  StreamInvocationMessage(String target, List<Object> arguments,
-      MessageHeaders headers, String invocationId)
+  StreamInvocationMessage(
+      {String? target,
+      List<Object>? arguments,
+      List<String>? streamIds,
+      MessageHeaders? headers,
+      String? invocationId})
       : this.target = target,
         this.arguments = arguments,
+        this.streamIds = streamIds,
         super(MessageType.StreamInvocation, headers, invocationId);
+
+  @override
+  String toString() {
+    return 'StreamInvocationMessage - type: ${type.index}, headers: ${headers.toString()}, invocationId: $invocationId, target: $target, arguments: $arguments, streamIds: $streamIds';
+  }
 }
 
 /// A hub message representing a single item produced as part of a result stream.
@@ -178,12 +204,18 @@ class StreamItemMessage extends HubInvocationMessage {
   // Properites
 
   /// The item produced by the server.
-  final Object item;
+  final Object? item;
 
   // Methods
-  StreamItemMessage(Object item, MessageHeaders headers, String invocationId)
+  StreamItemMessage(
+      {Object? item, MessageHeaders? headers, String? invocationId})
       : this.item = item,
         super(MessageType.StreamItem, headers, invocationId);
+
+  @override
+  String toString() {
+    return 'StreamInvocationMessage - type: ${type.index}, headers: ${headers.toString()}, invocationId: $invocationId, item: $item';
+  }
 }
 
 /// A hub message representing the result of an invocation.
@@ -193,19 +225,26 @@ class CompletionMessage extends HubInvocationMessage {
   /// The error produced by the invocation, if any.
   ///
   /// Either CompletionMessage.error CompletionMessage.result must be defined, but not both.
-  final String error;
+  final String? error;
 
   /// The result produced by the invocation, if any.
   ///
-  /// Either {@link @aspnet/signalr.CompletionMessage.error} or {@link @aspnet/signalr.CompletionMessage.result} must be defined, but not both.
-  final Object result;
+  /// Either {@link @microsoft/signalr.CompletionMessage.error} or {@link @microsoft/signalr.CompletionMessage.result} must be defined, but not both.
+  final Object? result;
 
   // Methods
   CompletionMessage(
-      String error, Object result, MessageHeaders headers, String invocationId)
+      {String? error,
+      Object? result,
+      MessageHeaders? headers,
+      String? invocationId})
       : this.error = error,
         this.result = result,
         super(MessageType.Completion, headers, invocationId);
+  @override
+  String toString() {
+    return 'CompletionMessage - type: ${type.index}, headers: ${headers.toString()}, invocationId: $invocationId, error: $error, result: $result';
+  }
 }
 
 /// A hub message indicating that the sender is still active.
@@ -213,11 +252,15 @@ class PingMessage extends HubMessageBase {
   // Methods
 
   PingMessage() : super(MessageType.Ping);
+  @override
+  String toString() {
+    return 'PingMessage - type: ${type.index}';
+  }
 }
 
 /// A hub message indicating that the sender is closing the connection.
 ///
-/// If {@link @aspnet/signalr.CloseMessage.error} is defined, the sender is closing the connection due to an error.
+/// If {@link @microsoft/signalr.CloseMessage.error} is defined, the sender is closing the connection due to an error.
 ///
 class CloseMessage extends HubMessageBase {
   // Properites
@@ -225,19 +268,33 @@ class CloseMessage extends HubMessageBase {
   /// The error that triggered the close, if any.
   ///
   /// If this property is undefined, the connection was closed normally and without error.
-  final String error;
+  final String? error;
+
+  /// If true, clients with automatic reconnects enabled should attempt to reconnect after receiving the CloseMessage. Otherwise, they should not. */
+  final bool? allowReconnect;
 
   //Methods
-  CloseMessage(String error)
+  CloseMessage({String? error, bool? allowReconnect})
       : this.error = error,
+        this.allowReconnect = allowReconnect,
         super(MessageType.Close);
+
+  @override
+  String toString() {
+    return 'CloseMessage - type: $type.index, allowReconnect: $allowReconnect, error: $error';
+  }
 }
 
 /// A hub message sent to request that a streaming invocation be canceled.
 class CancelInvocationMessage extends HubInvocationMessage {
   // Methods
-  CancelInvocationMessage(MessageHeaders headers, String invocationId)
+  CancelInvocationMessage({MessageHeaders? headers, String? invocationId})
       : super(MessageType.CancelInvocation, headers, invocationId);
+
+  @override
+  String toString() {
+    return 'CancelInvocationMessage - type: ${type.index}, headers: ${headers.toString()}, invocationId: $invocationId';
+  }
 }
 
 /// A protocol abstraction for communicating with SignalR Hubs.
@@ -259,14 +316,14 @@ abstract class IHubProtocol {
         this.version = number,
         this.transferFormat = transferFormat;
 
-  /// Creates an array of {@link @aspnet/signalr.HubMessage} objects from the specified serialized representation.
+  /// Creates an array of {@link @microsoft/signalr.HubMessage} objects from the specified serialized representation.
   ///
   /// If transferFormat is 'Text', the `input` parameter must be a string, otherwise it must be an ArrayBuffer.
   ///
   /// [input] A string (json), or Uint8List (binary) containing the serialized representation.
   /// [Logger] logger A logger that will be used to log messages that occur during parsing.
 
-  List<HubMessageBase> parseMessages(Object input, Logger logger);
+  List<HubMessageBase> parseMessages(Object input, Logger? logger);
 
   /// Writes the specified HubMessage to a string or ArrayBuffer and returns it.
   ///
