@@ -4,6 +4,8 @@ import 'errors.dart';
 
 /// Specifies a specific HTTP transport type.
 enum HttpTransportType {
+  /// Specified no transport preference. */
+  None, // = 0,
   /// Specifies the WebSockets transport. */
   WebSockets, // = 1,
   /// Specifies the Server-Sent Events transport. */
@@ -11,22 +13,20 @@ enum HttpTransportType {
   /// Specifies the Long Polling transport. */
   LongPolling, // = 4,
 }
-HttpTransportType httpTransportTypeFromString(String value) {
-  if (value == null) {
-    return null;
+
+HttpTransportType httpTransportTypeFromString(String? value) {
+  if (value == null || value == "") {
+    return HttpTransportType.None;
   }
 
   value = value.toUpperCase();
   switch (value) {
     case "WEBSOCKETS":
       return HttpTransportType.WebSockets;
-      break;
     case "SERVERSENTEVENTS":
       return HttpTransportType.ServerSentEvents;
-      break;
     case "LONGPOLLING":
       return HttpTransportType.LongPolling;
-      break;
     default:
       throw new GeneralError("$value is not a supported HttpTransportType");
   }
@@ -34,25 +34,25 @@ HttpTransportType httpTransportTypeFromString(String value) {
 
 /// Specifies the transfer format for a connection.
 enum TransferFormat {
+  /// TransferFormat is not defined.
+  Undefined, // = 0,
   /// Specifies that only text data will be transmitted over the connection.
   Text, // = 1,
   /// Specifies that binary data will be transmitted over the connection.
   Binary, // = 2,
 }
 
-TransferFormat getTransferFormatFromString(String value) {
-  if (value == null) {
-    return null;
+TransferFormat getTransferFormatFromString(String? value) {
+  if (value == null || value == "") {
+    return TransferFormat.Undefined;
   }
 
   value = value.toUpperCase();
   switch (value) {
     case "TEXT":
       return TransferFormat.Text;
-      break;
     case "BINARY":
       return TransferFormat.Binary;
-      break;
     default:
       throw new GeneralError("$value is not a supported HttpTransportType");
   }
@@ -60,20 +60,20 @@ TransferFormat getTransferFormatFromString(String value) {
 
 /// Data received call back.
 /// data: the content. Either a string (json) or Uint8List (binary)
-typedef OnReceive = void Function(Object data);
+typedef OnReceive = void Function(Object? data);
 
 ///
-typedef OnClose = void Function(Exception error);
+typedef OnClose = void Function({Exception? error});
 
 typedef AccessTokenFactory = Future<String> Function();
 
 /// An abstraction over the behavior of transports. This is designed to support the framework and not intended for use by applications.
 abstract class ITransport {
-  Future<void> connect(String url, TransferFormat transferFormat);
+  Future<void> connect(String? url, TransferFormat transferFormat);
 
   /// data: the content. Either a string (json) or Uint8List (binary)
   Future<void> send(Object data);
-  Future<void> stop(Error error);
-  OnReceive onReceive;
-  OnClose onClose;
+  Future<void> stop();
+  OnReceive? onReceive;
+  OnClose? onClose;
 }
