@@ -12,33 +12,32 @@ import 'utils.dart';
 class LongPollingTransport implements ITransport {
   // Properties
   final SignalRHttpClient _httpClient;
-  final AccessTokenFactory _accessTokenFactory;
-  final Logger _logger;
+  final AccessTokenFactory? _accessTokenFactory;
+  final Logger? _logger;
   final bool _logMessageContent;
   final AbortController _pollAbort;
 
-  bool get pollAborted => _pollAbort.aborted;
+  bool? get pollAborted => _pollAbort.aborted;
 
-  String _url;
-  bool _running;
-  Future<void> _receiving;
-  Exception _closeError;
-
-  @override
-  OnClose onClose;
+  String? _url;
+  late bool _running;
+  Future<void>? _receiving;
+  Exception? _closeError;
 
   @override
-  OnReceive onReceive;
+  OnClose? onClose;
+
+  @override
+  OnReceive? onReceive;
 
   // Methods
 
   LongPollingTransport(
       SignalRHttpClient httpClient,
-      AccessTokenFactory accessTokenFactory,
-      Logger logger,
+      AccessTokenFactory? accessTokenFactory,
+      Logger? logger,
       bool logMessageContent)
-      : assert(httpClient != null),
-        _httpClient = httpClient,
+      : _httpClient = httpClient,
         _accessTokenFactory = accessTokenFactory,
         _logger = logger,
         _logMessageContent = logMessageContent,
@@ -47,9 +46,8 @@ class LongPollingTransport implements ITransport {
   }
 
   @override
-  Future<void> connect(String url, TransferFormat transferFormat) async {
+  Future<void> connect(String? url, TransferFormat transferFormat) async {
     assert(!isStringEmpty(url));
-    assert(transferFormat != null);
 
     _url = url;
 
@@ -87,7 +85,7 @@ class LongPollingTransport implements ITransport {
     _receiving = poll(_url, pollOptions);
   }
 
-  Future<void> poll(String url, SignalRHttpRequest pollOptions) async {
+  Future<void> poll(String? url, SignalRHttpRequest pollOptions) async {
     try {
       while (_running) {
         // We have to get the access token on each poll, in case it changes
@@ -113,11 +111,11 @@ class LongPollingTransport implements ITransport {
             _running = false;
           } else {
             // Process the response
-            if (!isStringEmpty(response.content)) {
+            if (!isStringEmpty(response.content as String?)) {
               // _logger.log(LogLevel.Trace, "(LongPolling transport) data received. ${getDataDetail(response.content, this.logMessageContent)}");
               _logger?.finest("(LongPolling transport) data received");
               if (onReceive != null) {
-                onReceive(response.content);
+                onReceive!(response.content);
               }
             } else {
               // This is another way timeout manifest.
@@ -129,7 +127,7 @@ class LongPollingTransport implements ITransport {
           if (!_running) {
             // Log but disregard errors that occur after stopping
             _logger?.finest(
-                "(LongPolling transport) Poll errored after shutdown: ${e.message}");
+                "(LongPolling transport) Poll errored after shutdown: ${e.toString()}");
           } else {
             if (e is TimeoutError) {
               // Ignore timeouts and reissue the poll.
@@ -137,7 +135,7 @@ class LongPollingTransport implements ITransport {
                   "(LongPolling transport) Poll timed out, reissuing.");
             } else {
               // Close the connection with the error as the result.
-              _closeError = e;
+              _closeError = Exception(e.toString());
               _running = false;
             }
           }
@@ -148,7 +146,7 @@ class LongPollingTransport implements ITransport {
 
       // We will reach here with pollAborted==false when the server returned a response causing the transport to stop.
       // If pollAborted==true then client initiated the stop and the stop method will raise the close event after DELETE is sent.
-      if (!this.pollAborted) {
+      if (!this.pollAborted!) {
         _raiseOnClose();
       }
     }
@@ -165,7 +163,7 @@ class LongPollingTransport implements ITransport {
   }
 
   @override
-  Future<void> stop(Error error) async {
+  Future<void> stop() async {
     _logger?.finest("(LongPolling transport) Stopping polling.");
 
     // Tell receiving loop to stop, abort any current request, and then wait for it to finish
@@ -194,23 +192,23 @@ class LongPollingTransport implements ITransport {
     }
   }
 
-  Future<String> _getAccessToken() async {
+  Future<String?> _getAccessToken() async {
     if (_accessTokenFactory != null) {
-      return await _accessTokenFactory();
+      return await _accessTokenFactory!();
     }
     return null;
   }
 
-  void _updateHeaderToken(SignalRHttpRequest request, String token) {
+  void _updateHeaderToken(SignalRHttpRequest request, String? token) {
     if (request.headers == null) {
       request.headers = MessageHeaders();
     }
 
     if (!isStringEmpty(token)) {
-      request.headers.setHeaderValue("Authorization", "Bearer $token");
+      request.headers!.setHeaderValue("Authorization", "Bearer $token");
       return;
     }
-    request.headers.removeHeader("Authorization");
+    request.headers!.removeHeader("Authorization");
   }
 
   void _raiseOnClose() {
@@ -220,7 +218,7 @@ class LongPollingTransport implements ITransport {
         logMessage += " Error: " + _closeError.toString();
       }
       _logger?.finest(logMessage);
-      onClose(new GeneralError(_closeError?.toString()));
+      onClose!(error: GeneralError(_closeError?.toString()));
     }
   }
 }
