@@ -8,35 +8,35 @@ class SignalRHttpRequest {
   // Properties
 
   /// The HTTP method to use for the request.
-  String method;
+  String? method;
 
   /// The URL for the request.
-  String url;
+  String? url;
 
   ///The body content for the request. May be a string or an ArrayBuffer (for binary data).
-  Object content;
+  Object? content;
 
   ///An object describing headers to apply to the request.
-  MessageHeaders headers;
+  MessageHeaders? headers;
 
   // /** The XMLHttpRequestResponseType to apply to the request. */
   // responseType?: XMLHttpRequestResponseType;
 
   /// An AbortSignal that can be monitored for cancellation.
-  IAbortSignal abortSignal;
+  IAbortSignal? abortSignal;
 
   ///The time to wait for the request to complete before throwing a TimeoutError. Measured in milliseconds.
-  int timeout;
+  int? timeout;
 
   // Methods
 
   SignalRHttpRequest(
-      {String method,
-      String url,
-      Object content,
-      MessageHeaders headers,
-      IAbortSignal abortSignal,
-      int timeout})
+      {String? method,
+      String? url,
+      Object? content,
+      MessageHeaders? headers,
+      IAbortSignal? abortSignal,
+      int? timeout})
       : this.method = method,
         this.url = url,
         this.content = content,
@@ -53,10 +53,10 @@ class SignalRHttpResponse {
   final int statusCode;
 
   /// The status message of the response
-  final String statusText;
+  final String? statusText;
 
   /// May be a string (json) or an Uint8List (binary)
-  final Object content;
+  final Object? content;
 
   //Methods
 
@@ -66,7 +66,8 @@ class SignalRHttpResponse {
   /// statusText: The status message of the response.
   /// content: The content of the response
   ///
-  SignalRHttpResponse(int statusCode, {String statusText = '', Object content})
+  SignalRHttpResponse(int statusCode,
+      {String? statusText = '', Object? content})
       : this.statusCode = statusCode,
         this.statusText = statusText,
         this.content = content;
@@ -77,37 +78,40 @@ class SignalRHttpResponse {
 /// This class provides an abstraction over an HTTP client so that a different implementation can be provided on different platforms.
 ///
 abstract class SignalRHttpClient {
-  /// Issues an HTTP GET request to the specified URL, returning a Promise that resolves with an {@link @aspnet/signalr.HttpResponse} representing the result.
+  /// Issues an HTTP GET request to the specified URL, returning a Promise that resolves with an {@link @microsoft/signalr.HttpResponse} representing the result.
   ///
   /// url The URL for the request.
   /// HttpRequest options Additional options to configure the request. The 'url' field in this object will be overridden by the url parameter.
   /// Returns a Future<HttpResponse> that resolves with an HttpResponse describing the response, or rejects with an Error indicating a failure.
   ///
-  Future<SignalRHttpResponse> get(String url, {SignalRHttpRequest options}) {
+  Future<SignalRHttpResponse> get(String url,
+      {required SignalRHttpRequest options}) {
     options.method = 'GET';
     options.url = url;
     return send(options);
   }
 
-  /// Issues an HTTP POST request to the specified URL, returning a Promise that resolves with an {@link @aspnet/signalr.HttpResponse} representing the result.
+  /// Issues an HTTP POST request to the specified URL, returning a Promise that resolves with an {@link @microsoft/signalr.HttpResponse} representing the result.
   ///
   /// url: The URL for the request.
   /// options: Additional options to configure the request. The 'url' field in this object will be overridden by the url parameter.
   /// Returns a Future that resolves with an describing the response, or rejects with an Error indicating a failure.
   ///
-  Future<SignalRHttpResponse> post(String url, {SignalRHttpRequest options}) {
+  Future<SignalRHttpResponse> post(String? url,
+      {required SignalRHttpRequest options}) {
     options.method = 'POST';
     options.url = url;
     return send(options);
   }
 
-  ///Issues an HTTP DELETE request to the specified URL, returning a Promise that resolves with an {@link @aspnet/signalr.HttpResponse} representing the result.
+  ///Issues an HTTP DELETE request to the specified URL, returning a Promise that resolves with an {@link @microsoft/signalr.HttpResponse} representing the result.
   ///
   /// The URL for the request.
   /// Additional options to configure the request. The 'url' field in this object will be overridden by the url parameter.
   /// Returns a Future that resolves with an describing the response, or rejects with an Error indicating a failure.
   ///
-  Future<SignalRHttpResponse> delete(String url, {SignalRHttpRequest options}) {
+  Future<SignalRHttpResponse> delete(String? url,
+      {required SignalRHttpRequest options}) {
     options.method = 'DELETE';
     options.url = url;
     return send(options);
