@@ -1,6 +1,5 @@
-import 'package:client/utils/viewModel/viewModel.dart';
+import 'viewModel.dart';
 import 'package:flutter/widgets.dart';
-
 
 class ViewModelProvider<TViewModel extends ViewModel> extends InheritedWidget {
   // Properties
@@ -8,12 +7,20 @@ class ViewModelProvider<TViewModel extends ViewModel> extends InheritedWidget {
   final TViewModel viewModel;
   // Methods
 
-  ViewModelProvider({Key key, @required TViewModel viewModel, @required WidgetBuilder childBuilder}) : this._default(key: key, viewModel: viewModel, childBuilder: childBuilder);
+  ViewModelProvider(
+      {Key? key,
+      required TViewModel? viewModel,
+      required WidgetBuilder? childBuilder})
+      : this._default(
+            key: key, viewModel: viewModel, childBuilder: childBuilder);
 
-  ViewModelProvider._default({Key key, @required TViewModel viewModel, @required WidgetBuilder childBuilder})
+  ViewModelProvider._default(
+      {Key? key,
+      required TViewModel? viewModel,
+      required WidgetBuilder? childBuilder})
       : assert(viewModel != null),
         assert(childBuilder != null),
-        viewModel = viewModel,
+        viewModel = viewModel!,
         super(
             key: key,
             child: ViewModelViewStateManager(
@@ -34,11 +41,12 @@ class ViewModelViewStateManager extends StatefulWidget {
 
   // Methods
 
-  ViewModelViewStateManager({@required ViewModel viewModel, @required WidgetBuilder childBuilder})
+  ViewModelViewStateManager(
+      {required ViewModel? viewModel, required WidgetBuilder? childBuilder})
       : assert(viewModel != null),
         assert(childBuilder != null),
-        _childBuilder = childBuilder,
-        _viewModel = viewModel;
+        _childBuilder = childBuilder!,
+        _viewModel = viewModel!;
 
   @override
   State<StatefulWidget> createState() => _ViewModelViewStateManagerState();
