@@ -1,32 +1,34 @@
 import 'dart:async';
 
-import 'package:client/main.dart';
-import 'package:client/tests/tests.dart';
-import 'package:client/utils/viewModel/viewModel.dart';
-import 'package:client/utils/viewModel/viewModelProvider.dart';
-import 'package:flutter/foundation.dart';
+import '../../main.dart';
+import '../../tests/tests.dart';
+import '../../utils/viewModel/viewModel.dart';
+import '../../utils/viewModel/viewModelProvider.dart';
 import 'package:flutter/widgets.dart';
 import 'package:logging/logging.dart';
-import 'package:signalr_client/signalr_client.dart';
+import 'package:signalr_netcore/ihub_protocol.dart';
+//import 'package:signalr_netcore/msgpack_hub_protocol.dart';
+import 'package:signalr_netcore/signalr_client.dart';
 
 typedef HubConnectionProvider = Future<HubConnection> Function();
 
 class TestsPageViewModel extends ViewModel {
 // Properties
-  Logger _logger;
-  StreamSubscription<LogRecord> _logMessagesSub;
-  Tests _tests;
-  String _serverUrl;
-  HubConnection _hubConnection;
+  late Logger _logger;
+  late StreamSubscription<LogRecord> _logMessagesSub;
+  late Tests _tests;
+  late String _serverUrl;
+  HubConnection? _hubConnection;
 
-  String _errorMessage;
+  late String _errorMessage;
   static const String errorMessagePropName = "errorMessage";
   String get errorMessage => _errorMessage;
   set errorMessage(String value) {
-    updateValue(errorMessagePropName, _errorMessage, value, (v) => _errorMessage = v);
+    updateValue<String>(
+        errorMessagePropName, _errorMessage, value, (v) => _errorMessage = v);
   }
 
-  List<LogRecord> _hubLogMessages;
+  late List<LogRecord> _hubLogMessages;
   static const String hubLogMessagesPropName = "hubLogMessages";
   List<LogRecord> get hubLogMessages => _hubLogMessages;
 
@@ -34,9 +36,9 @@ class TestsPageViewModel extends ViewModel {
 
 // Methods
   TestsPageViewModel() {
-    _hubLogMessages = List<LogRecord>();
+    _hubLogMessages = [];
 
-    Logger.root.level = Level.FINEST;
+    Logger.root.level = Level.ALL;
     _logMessagesSub = Logger.root.onRecord.listen(_handleLogMessage);
     _logger = Logger("TestsPageViewModel");
 
@@ -51,28 +53,38 @@ class TestsPageViewModel extends ViewModel {
   }
 
   void _handleLogMessage(LogRecord msg) {
+    //print(msg);
     _hubLogMessages.add(msg);
     notifyPropertyChanged(hubLogMessagesPropName);
   }
 
   Future<HubConnection> _getHubConnection() async {
-    //final logger = _logger;
-    final logger = null;
+    final logger = _logger;
+    //final logger = null;
     if (_hubConnection == null) {
-      final httpOptions = new HttpConnectionOptions(logger: logger);
+      final headers = MessageHeaders();
+      headers.setHeaderValue("api-key", "my-top-secret-api-key");
+      final httpOptions =
+          new HttpConnectionOptions(logger: logger, headers: headers);
       //final httpOptions = new HttpConnectionOptions(logger: logger, transport: HttpTransportType.ServerSentEvents);
       //final httpOptions = new HttpConnectionOptions(logger: logger, transport: HttpTransportType.LongPolling);
 
-      _hubConnection = HubConnectionBuilder().withUrl(_serverUrl, options: httpOptions).configureLogging(logger).build();
-      _hubConnection.onclose((error) => _logger.info("Connection Closed"));
+      _hubConnection = HubConnectionBuilder()
+          .withUrl(_serverUrl, options: httpOptions)
+          /* Configure the Hub with msgpack protocol */
+          //.withHubProtocol(MessagePackHubProtocol())
+          .withAutomaticReconnect()
+          .configureLogging(logger)
+          .build();
+      _hubConnection!.onclose(({error}) => _logger.info("Connection Closed"));
     }
 
-    if (_hubConnection.state != HubConnectionState.Connected) {
-      await _hubConnection.start();
-      _logger.info("Connection state '${_hubConnection.state}'");
+    if (_hubConnection!.state != HubConnectionState.Connected) {
+      await _hubConnection!.start();
+      _logger.info("Connection state '${_hubConnection!.state}'");
     }
 
-    return _hubConnection;
+    return _hubConnection!;
   }
 
   Future<void> connect() async {
@@ -91,9 +103,13 @@ class TestsPageViewModelProvider extends ViewModelProvider<TestsPageViewModel> {
   // Properties
 
   // Methods
-  TestsPageViewModelProvider({Key key, viewModel: TestsPageViewModel, WidgetBuilder childBuilder}) : super(key: key, viewModel: viewModel, childBuilder: childBuilder);
+  TestsPageViewModelProvider(
+      {Key? key, viewModel = TestsPageViewModel, WidgetBuilder? childBuilder})
+      : super(key: key, viewModel: viewModel, childBuilder: childBuilder);
 
-  static TestsPageViewModel of(BuildContext context) {
-    return (context.inheritFromWidgetOfExactType(TestsPageViewModelProvider) as TestsPageViewModelProvider).viewModel;
+  static TestsPageViewModel? of(BuildContext context) {
+    return context
+        .dependOnInheritedWidgetOfExactType<TestsPageViewModelProvider>()
+        ?.viewModel;
   }
 }
