@@ -1,6 +1,6 @@
 library signalr_client;
 
-export 'dartio_http_client.dart';
+export 'web_supporting_http_client.dart';
 export 'http_connection.dart';
 export 'http_connection_options.dart';
 export 'hub_connection.dart';
