@@ -1,13 +1,16 @@
-import 'package:client/tests/test.dart';
-import 'package:client/views/pages/testsPageViewModel.dart';
+import '../../views/pages/testsPageViewModel.dart';
+import '../test.dart';
 import 'package:logging/logging.dart';
-import 'package:signalr_client/signalr_client.dart';
+import 'package:signalr_netcore/signalr_client.dart';
 
 class ClientMethodNoParametersNoReturnValue extends Test {
   // Properties
 
   // Methods
-  ClientMethodNoParametersNoReturnValue(HubConnectionProvider hubConnectionProvider, Logger logger) : super(hubConnectionProvider, logger, "Client Invokes method: 'MethodNoParametersNoReturnValue");
+  ClientMethodNoParametersNoReturnValue(
+      HubConnectionProvider hubConnectionProvider, Logger logger)
+      : super(hubConnectionProvider, logger,
+            "Client Invokes method: 'MethodNoParametersNoReturnValue");
 
   @override
   Future<void> executeTest(HubConnection hubConnection) async {
