@@ -1,17 +1,20 @@
-import 'package:client/tests/test.dart';
-import 'package:client/views/pages/testsPageViewModel.dart';
+import '../../views/pages/testsPageViewModel.dart';
+import '../test.dart';
 import 'package:logging/logging.dart';
-import 'package:signalr_client/signalr_client.dart';
+import 'package:signalr_netcore/signalr_client.dart';
 
 class ClientMethodOneSimpleParameterNoReturnValue extends Test {
   // Properties
 
   // Methods
-  ClientMethodOneSimpleParameterNoReturnValue(HubConnectionProvider hubConnectionProvider, Logger logger)
-      : super(hubConnectionProvider, logger, "Client Invokes method 'MethodOneSimpleParameterNoReturnValue");
+  ClientMethodOneSimpleParameterNoReturnValue(
+      HubConnectionProvider hubConnectionProvider, Logger logger)
+      : super(hubConnectionProvider, logger,
+            "Client Invokes method 'MethodOneSimpleParameterNoReturnValue");
 
   @override
   Future<void> executeTest(HubConnection hubConnection) async {
-    await hubConnection.invoke("MethodOneSimpleParameterNoReturnValue", args: <Object>["ParameterValue"]);
+    await hubConnection.invoke("MethodOneSimpleParameterNoReturnValue",
+        args: <Object>["ParameterValue"]);
   }
 }
