@@ -1,18 +1,22 @@
-import 'package:client/tests/test.dart';
-import 'package:client/views/pages/testsPageViewModel.dart';
+import '../../views/pages/testsPageViewModel.dart';
+import '../test.dart';
 import 'package:logging/logging.dart';
-import 'package:signalr_client/signalr_client.dart';
+import 'package:signalr_netcore/signalr_client.dart';
 
 class ClientMethodOneSimpleParameterSimpleReturnValue extends Test {
   // Properties
 
   // Methods
-  ClientMethodOneSimpleParameterSimpleReturnValue(HubConnectionProvider hubConnectionProvider, Logger logger)
-      : super(hubConnectionProvider, logger, "Client Invokes method 'MethodOneSimpleParameterSimpleReturnValue");
+  ClientMethodOneSimpleParameterSimpleReturnValue(
+      HubConnectionProvider hubConnectionProvider, Logger logger)
+      : super(hubConnectionProvider, logger,
+            "Client Invokes method 'MethodOneSimpleParameterSimpleReturnValue");
 
   @override
   Future<void> executeTest(HubConnection hubConnection) async {
-    final result = await hubConnection.invoke("MethodOneSimpleParameterSimpleReturnValue", args: <Object>["ParameterValue"]);
+    final result = await hubConnection.invoke(
+        "MethodOneSimpleParameterSimpleReturnValue",
+        args: <Object>["ParameterValue"]);
     logger.info("Result: '$result");
   }
 }
